pyproject.toml
[tool.poetry]
name = "oc_prompt_study"
version = "0.1.0"
description = ""
authors = ["Hyeongsik Kim <Hyeongsik.Kim@us.bosch.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.14,>=3.10"
langchain = "^0.3.15"
langchain-openai = "^0.3.1"
langchain-azure-ai = "^0.1.2"
langchain-community = "^0.3.15"
deepeval = "^3.4.8"
langchain-ollama = "^0.2.2"
ollama = "^0.4.7"
pydantic = "^2.11.4"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
beautifulsoup4 = "^4.12.3"
scipy = "^1.15.2"
Jinja2 = "^3.1.4"
openpyxl = "^3.1.5"
numpy = "^1.26.4"
pingouin = "^0.5.5"
statsmodels = "^0.14.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
